02_introduction_Agno/requirements.txt
agno
python-dotenv
groq
duckduckgo-search
